+++ src/main/java/com/froggylord/constellation/mixin/ItemProtectionScreenMixin.java
@@ -65,7 +65,8 @@ public abstract class ItemProtectionScreenMixin {
     @Inject(method = "getTooltipFromContainerItem", at = @At("RETURN"), cancellable = true)
     private void constellation$protectedTooltip(ItemStack stack, CallbackInfoReturnable<List<Component>> cir) {
         AbstractContainerScreen<?> screen = (AbstractContainerScreen<?>) (Object) this;
-        cir.setReturnValue(com.froggylord.constellation.constellation.HerculesPlotPrices.appendTooltip(screen, stack,
+        cir.setReturnValue(com.froggylord.constellation.constellation.HerculesCarrolynHelper.appendTooltip(screen, stack,
+            com.froggylord.constellation.constellation.HerculesPlotPrices.appendTooltip(screen, stack,
             com.froggylord.constellation.constellation.HerculesPesthunterShop.appendTooltip(screen, stack,
             com.froggylord.constellation.constellation.HerculesVisitorLogbook.appendTooltip(screen, stack,
             com.froggylord.constellation.constellation.HerculesAnitaShop.appendTooltip(screen, stack,
@@ -79,6 +80,6 @@ public abstract class ItemProtectionScreenMixin {
             com.froggylord.constellation.constellation.HerculesVisitorHelper.appendTooltip(screen, stack,
                 com.froggylord.constellation.constellation.LyraAuctionHelper.appendTooltip(screen, stack,
                     com.froggylord.constellation.constellation.LyraBazaarHelper.appendTooltip(screen, stack,
-                        ItemProtection.appendTooltip(stack, cir.getReturnValue()))))))))))))))));
+                        ItemProtection.appendTooltip(stack, cir.getReturnValue())))))))))))))))));
     }
 }